refactor: yearly patterns
