+++ components/RecurringTasks/streaks/patterns/yearly/types.ts
@@ -1,0 +1,8 @@
+import { DateRange } from '../../types';
+
+export interface YearlyTarget {
+  date: Date;
+  allowedRange: DateRange;
+  month: number;
+  dayOfMonth: number;
+}
